requirements.txt
appdirs==1.4.3
configparser==3.5.0
Django==1.11.1
djangorestframework==3.6.3
mysqlclient==1.3.10
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
